+++ ChangeLog.md
@@ -7,11 +7,13 @@
 
 - Added IntValue, FloatValue and StringValue as specialized containers for scalars and strings, to be used in place of BaseValue where needed.
 - Added Phasescreen data object
+- Added VideoRecorder processing object
 
 ### Interface changes
 
 - Removed simul\_params argument from IirFilter, Integrator and other related processing objects
 - Outputs for SpeculaInput and derived objects like TerminalInput must be typed with :int, :float or :str
+- Added "window" and "subplot" arguments to all display to enable multi-plot windows
 
 ### Other